list with cards first draft complete
